More-excellent font!
Includes codepoints for control characters like CAPS SHIFT, and arrow
keys, which have graphical representations.

This helps with debugging keyboard decode logic, but also the glyphs
can be handy for representing CAPS LOCK state, and cardinal
directions in programs.

Slight issue in that some characters are 8 pixels high… which is not
currently supported in the font encoding. (It will be eventually
though.)