Fix (existing) bug in makefont which was including spurious rows in glyphs
Some glyphs (I noticed ‘E’ and ‘F’, but there might have been others)
were getting blank initial and trailing lines included in the pixel
data.

The font was a few bytes larger than it should have been, and
rendering of the letter ‘E’ and ‘F’ was a few nanoseconds sluggish.