Remove complexity, size & slowness from UDG-copying
The previous code was general enough to handle any valid glyph header.
However, we’re only ever copying glyphs 6 pixels tall with 1 px
leading, so we can specialise (hardcode) the copying code to that.

There’s a bit of logic in genfont and the assembly code to verify
(assert) this assumption at built-time.

Went from 60 bytes to 26 bytes of copying code.