Faster rendering of offset, 8-pixel-width character
Simpler AND faster! (And taking less memory!)

We had been doing memory lookups to rotate bits when rendering
1-cell-width chars by 4 pixels, but it’s strictly faster to just do
4 rrca operations, then mask out for left and right side. So get rid
of the lookup tables too (512 bytes saved)!

And also means no setup (exx; ld h,<constant>; exx), and because we
don’t clobber `A'`, the scoot back up to the top of the character
cell at the end of rendering the glyph is also faster.