All ‘long’ tokens have a 1-character form
‘Long’ tokens at tokenisation stage (identifiers, numeric literals,
comments, etc) have a 1-character form, which is shorter. We were
already doing that for identifiers, but generalise it to all long
tokens. (Any token with a delimited will in practice not actually get
there, so it only applies to 1-char identifiers, 1-digit decimal
literals, completely empty comments, and 1-char invalid tokens.

* Reduces the capacity limit for tokenised line to not (much) more than
  2x the raw characters. (Missing spaces between tokens can inflate
  that.)
* Regularises the code.