Token stream always < code buffer bytes * 2
Enforce invariant that size of the token buffer need be no more than
twice the code character buffer in length.

Do this by encoding the "missing space" condition in the bits of the
subsequent token.

We ‘remember’ a missing space in the carry bit of AF'.

Slightly chonkier code, but not awful. Adds ~16 tstates per token.
(Though we avoid writing the "NOSPACE" token ever.)